refactor: consolidate error handling into a single error module
