+++ src/market_data_ingestor/src/error.rs
